+++ crates/jmap/src/services/ingest.rs
@@ -34,7 +34,7 @@ impl JMAP {
         // Read message
         let raw_message = match self
             .blob_store
-            .get_blob(message.message_blob.as_slice(), 0..u32::MAX)
+            .get_blob(message.message_blob.as_slice(), 0..usize::MAX)
             .await
         {
             Ok(Some(raw_message)) => raw_message,
